requirements.txt
PyMuPDF==1.22.3
Flask==2.3.2
numpy==1.21.0
pandas==1.3.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==9.5.0
Werkzeug==2.3.3